AI project risk chain/llm/llm/Supply-Chain-Risk-Analyzer/k/requirements.txt
PyQt6
mss
pynput
Pillow
requests
python-dotenv
